# openapi.yaml
openapi: 3.1.0
info:
  title: BearStack HTTP API
  version: 0.24.0
  description: |
    Programmatic HTTP interface for documents, uploads, tags, custom fields,
    virtual folders and document media.

    BearStack accepts HTTP Basic authentication with the configured credentials.
    A browser session cookie obtained through the web login is accepted as well.
    When authentication is disabled in the server configuration, the security
    requirements are not enforced.

    Write operations require the capabilities stated in their descriptions.
    API clients using form endpoints should send `Accept: application/json`.
servers:
  - url: /
    description: Current BearStack instance
security:
  - basicAuth: []
  - sessionCookie: []
...
paths:
...
  /api/upload:
    post:
      tags: [Upload]
      operationId: uploadDocuments
      summary: Upload one or more documents
      description: |
        Imports every file part as a document. Files exceeding the configured
        upload limit, unsupported file types and duplicates are reported per
        item. Requires the `documents.upload` capability.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    format: binary
            encoding:
              files:
                style: form
                explode: true
      responses:
        "201":
          description: |
            The request was processed and at least one file was imported, was a
            duplicate, or did not produce a per-file error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: No document was imported and at least one error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/fields:
    get:
      tags: [Structure]
      operationId: listCustomFields
      summary: List custom field definitions
      description: Requires the `documents.read` capability.
      responses:
        "200":
          description: Custom field definitions in display order.
          content:
            application/json:
              schema:
                type: object
                required: [fields]
                properties:
                  fields:
                    type: array
                    items:
                      $ref: "#/components/schemas/CustomField"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/tags:
    get:
      tags: [Structure]
      operationId: listTags
      summary: List document tags
      description: Requires the `documents.read` capability.
      responses:
        "200":
          description: All document tags.
          content:
            application/json:
              schema:
                type: object
                required: [tags]
                properties:
                  tags:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tag"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /documents/{id}/tags:
    post:
      tags: [Documents]
      operationId: replaceDocumentTags
      summary: Replace all tags on a document
      description: |
        Existing tags can be assigned with `documents.edit`. Creating a new tag
        additionally requires `documents.structure`. Deleted documents cannot be
        changed. Send `Accept: application/json`.
      parameters:
        - $ref: "#/components/parameters/DocumentID"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
            encoding:
              tags:
                style: form
                explode: true
      responses:
        "200":
          description: Normalized, stored tag names.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentTags"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /documents/{id}/metadata:
    post:
      tags: [Documents]
      operationId: updateDocumentMetadata
      summary: Update document metadata
      description: |
        Replaces title, description, date, tags and custom field values.
        Custom fields use form field names `field_{id}`, for example `field_3`.
        Existing tags can be assigned with `documents.edit`; creating a new tag
        additionally requires `documents.structure`. Send
        `Accept: application/json`.
      parameters:
        - $ref: "#/components/parameters/DocumentID"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                document_date:
                  description: Empty to clear the date.
                  oneOf:
                    - type: string
                      format: date
                    - type: string
                      maxLength: 0
                tags:
                  type: array
                  items:
                    type: string
                new_tags:
                  type: string
                  description: Additional comma-separated tags.
                field_{id}:
                  type: string
                  description: |
                    Custom field value. Replace `{id}` in the form field name
                    with a positive custom field ID.
              additionalProperties:
                type: string
            encoding:
              tags:
                style: form
                explode: true
      responses:
        "200":
          description: Selected values from the updated metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataUpdateResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /documents/{id}/document-date:
    post:
      tags: [Documents]
      operationId: updateDocumentDate
      summary: Update only the document date
      description: Requires `documents.edit` and always returns JSON.
      parameters:
        - $ref: "#/components/parameters/DocumentID"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                document_date:
                  description: Empty to clear the date.
                  oneOf:
                    - type: string
                      format: date
                    - type: string
                      maxLength: 0
      responses:
        "200":
          description: Formatted updated date.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDateUpdateResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /photos/people/{id}/rename:
    post:
      tags:
      - Photos
      operationId: renamePhotoPerson
      summary: Name or rename a person
      description: Requires photos.manage. Form writes redirect to the corresponding
        HTML page. Existing authentication, same-origin protection and audit logging
        apply.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 200
              required: []
      responses:
        '303':
          description: Saved; Location points to the corresponding page.
          headers:
            Location:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /photos/people/{id}/merge:
    post:
      tags:
      - Photos
      operationId: mergePhotoPeople
      summary: Merge source group into target
      description: Requires photos.manage. Form writes redirect to the corresponding
        HTML page. Existing authentication, same-origin protection and audit logging
        apply.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                target:
                  type: integer
                  format: int64
                  minimum: 1
              required:
              - target
      responses:
        '303':
          description: Saved; Location points to the corresponding page.
          headers:
            Location:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /photos/faces/edit:
    post:
      tags:
      - Photos
      operationId: editPhotoFaces
      summary: Move selected faces or ignore false detections
      description: Requires photos.manage. Form writes redirect to the corresponding
        HTML page. Existing authentication, same-origin protection and audit logging
        apply.
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                face_id:
                  type: array
                  items:
                    type: integer
                    format: int64
                    minimum: 1
                  minItems: 1
                  maxItems: 500
                action:
                  type: string
                  enum:
                  - move
                  - ignore
                target:
                  type: integer
                  minimum: 0
                  default: 0
                  description: Existing target person ID; 0 or omitted creates a new
                    person for move.
                name:
                  type: string
                  maxLength: 200
                  description: Optional name for the new person.
              required:
              - face_id
              - action
      responses:
        '303':
          description: Saved; Location points to the corresponding page.
          headers:
            Location:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /settings/photos/faces:
    get:
      tags:
      - Photos
      operationId: getFaceRecognitionSettings
      summary: Recognition settings and progress
      description: Requires photos.manage. HTML by default; format=json returns settings
        and status. Service tokens and vectors are never returned.
      parameters:
      - name: format
        in: query
        schema:
          type: string
          enum:
          - json
        description: Request JSON instead of HTML.
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceSettingsStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Photos
      operationId: saveFaceRecognitionSettings
      summary: Configure optional face recognition
      description: Requires photos.manage. Form writes redirect to the corresponding
        HTML page. Existing authentication, same-origin protection and audit logging
        apply.
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                enabled:
                  type: string
                  enum:
                  - '1'
                  description: Omit to pause. Activation requires a healthy, compatible
                    configured service.
                batch_size:
                  type: integer
                  minimum: 1
                  maximum: 1000
                  default: 100
                delay_millis:
                  type: integer
                  minimum: 100
                  maximum: 60000
                  default: 1000
                interval_minutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
                  default: 15
              required: []
      responses:
        '303':
          description: Saved; Location points to the corresponding page.
          headers:
            Location:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /settings/photos/faces/{action}:
    post:
      tags:
      - Photos
      operationId: controlFaceRecognition
      summary: Pause, resume, retry or delete recognition data
      description: Requires photos.manage. Form writes redirect to the corresponding
        HTML page. Existing authentication, same-origin protection and audit logging
        apply.
      parameters:
      - name: action
        in: path
        required: true
        schema:
          type: string
          enum:
          - pause
          - resume
          - retry
          - clear
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                confirm:
                  type: string
                  enum:
                  - delete
                  description: Required for clear; deletes generated data and disables
                    processing, preserving XMP.
              required: []
      responses:
        '303':
          description: Saved; Location points to the corresponding page.
          headers:
            Location:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Credentials configured for the BearStack instance.
    sessionCookie:
      type: apiKey
      in: cookie
      name: bearstack_session
      description: Signed session cookie created by the web login.
  parameters:
    DocumentID:
      name: id
      in: path
      required: true
      description: Positive document ID.
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Basic realm="BearStack", charset="UTF-8"
      content:
        text/plain:
          schema:
            type: string
    Forbidden:
      description: The authenticated account lacks the required capability.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalServerError:
      description: Internal error with implementation details masked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Interner Serverfehler
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UploadResult:
      type: object
      required: [uploaded, duplicates, errors]
      properties:
        uploaded:
          type: array
          items:
            $ref: "#/components/schemas/UploadedDocument"
        duplicates:
          type: array
          items:
            $ref: "#/components/schemas/DuplicateDocument"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/UploadError"
...
    CustomField:
      type: object
      required:
        - id
        - label
        - position
        - autocomplete_enabled
        - value_folder_min_documents
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        label:
          type: string
        position:
          type: integer
          minimum: 0
        autocomplete_enabled:
          type: boolean
        value_folder_min_documents:
          description: |
            `0` disables value folders, `1` always shows them, and higher
            values are the minimum matching document count.
          type: integer
          enum: [0, 1, 5, 10, 20, 50]
...
    Tag:
      type: object
      required:
        - id
        - name
        - display_name
        - description
        - color
        - primary_tag
        - group_mode
        - list_hidden
        - delete_protected
        - style
        - count
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        color:
          type: string
        primary_tag:
          type: boolean
        group_mode:
          type: boolean
        list_hidden:
          type: boolean
        delete_protected:
          type: boolean
        style:
          type: string
        count:
          type: integer
          minimum: 0
...